+++ src/api/auth.ts
@@ -1,7 +1,8 @@
-import axiosInstance from "./axiosInstance";
+import baseApi from "./Instance/baseApi";
+import authApi from "./Instance/authApi";
 
 export const login = async (studentId: string, password: string) => {
-  const response = await axiosInstance.post(`/auth/login`, {
+  const response = await baseApi.post(`/auth/login`, {
     studentId,
     password,
   });
@@ -9,5 +10,5 @@ export const login = async (studentId: string, password: string) => {
 };
 
 export const logout = async () => {
-  return await axiosInstance.post('/auth/logout');
-}
+  return await authApi.post("/auth/logout");
+};
